Done lab 03

+++ Source/Shape.cpp
@@ -1,6 +1,6 @@
 #include "../Header/Shape.h"
 
-Shape::Shape(int type, int x1, int y1) {
+Shape::Shape(int type, float x1, float y1) {
 	this->type = type;
 	this->x1 = x1;
 	this->y1 = y1;
